+++ lib/interpreter/instantiate/module.cpp
@@ -152,7 +152,8 @@ Expect<void> Interpreter::instantiate(Runtime::StoreManager &StoreMgr,
           .value_or(nullptr);
 
   ModInst->GlobalsPtr.reserve(ModInst->getGlobalNum());
-  for (size_t I = 0; I < ModInst->getGlobalNum(); ++I) {
+  for (uint32_t I = 0; I < static_cast<uint32_t>(ModInst->getGlobalNum());
+       ++I) {
     ModInst->GlobalsPtr.push_back(
         &(*StoreMgr.getGlobal(*ModInst->getGlobalAddr(I)))->getValue());
   }
